Fix QA issues.
