+++ src/databaseTest/java/net/sf/jabref/shared/TestManager.java
@@ -5,6 +5,10 @@ import java.util.Collection;
 import java.util.HashSet;
 import java.util.Set;
 
+/**
+ * {@link TestManager} provides helping methods for database tests.
+ * Furthermore it determines database systems which are ready to be used for tests.
+ */
 public class TestManager {
 
     public static Collection<DBMSType> getDBMSTypeTestParameter() {
