+++ src/main/resources/l10n/JabRef_de.properties
@@ -2252,7 +2252,7 @@ Local_version\:_%0=Lokale_Version\:_%0
 Shared_version\:_%0=Geteilte_genutzte_Version\:_%0
 Please_merge_the_shared_entry_with_yours_and_press_"Merge_entries"_to_resolve_this_problem.=Bitte_führen_Sie_den_gemeinsam_genutzten_Eintrag_mit_Ihrem_zusammen_und_klicken_Sie_auf_"Einträge_zuammenführen",_um_das_Problem_zu_beheben.
 Canceling_this_operation_will_leave_your_changes_unsynchronized._Cancel_anyway?=Durch_einen_Abbruch_dieser_Operation_werden_die_Änderungen_nicht_synchronisiert._Trotzdem_abbrechen?
-The_BibEntry_you_currently_work_on_has_been_deleted_on_the_shared_side.=Der_Eintrag,_welchen_Sie_zur_Zeit_bearbeiten,_ist_auf_der_gemeinsam_genutzen_Datenbank_nicht_mehr_vorhanden.
+The_BibEntry_you_currently_work_on_has_been_deleted_on_the_shared_side.=Der_Eintrag,_welchen_Sie_zur_Zeit_bearbeiten,_ist_auf_der_geteilten_Datenbank_nicht_mehr_vorhanden.
 Remember_password?=Passwort_merken?
 Cannot_cite_entries_without_BibTeX_keys._Generate_keys_now?=Kann_keine_Einträge_ohne_BibTeX-Key_zitieren._Keys_jetzt_generieren?
 New_technical_report=Neuer_technischer_Bericht
@@ -2306,5 +2306,5 @@ Search_in_all_open_databases=
 
 Invalid_ISBN\:_'%0'.=
 
-Shared_entry_is_no_longer_present=Gemeinsam_genutzter_Eintrag_ist_nicht_mehr_vorhanden
+Shared_entry_is_no_longer_present=Geteilter_Eintrag_ist_nicht_mehr_vorhanden
 You_can_restore_the_entry_using_the_"Undo"_operation.=Unter_Benutzung_der_"Rückgängig"-Funktion_kann_der_Eintrag_wiederhergestellt_werden.
