+++ src/oracleMock/oracle/jdbc/OracleConnection.java
@@ -22,7 +22,7 @@ import java.util.concurrent.Executor;
 import oracle.jdbc.dcn.DatabaseChangeRegistration;
 
 /**
- * Mocking class
+ * A mocking class used as a placeholder for the real Oracle JDBC drivers to prevent build errors.
  */
 public class OracleConnection implements Connection {
