+++ src/oracleMock/oracle/jdbc/OracleStatement.java
@@ -3,7 +3,7 @@ package oracle.jdbc;
 import oracle.jdbc.dcn.DatabaseChangeRegistration;
 
 /**
- * Mocking class
+ * A mocking class used as a placeholder for the real Oracle JDBC drivers to prevent build errors.
  */
 public class OracleStatement {
