+++ src/oracleMock/oracle/jdbc/dcn/DatabaseChangeEvent.java
@@ -1,7 +1,7 @@
 package oracle.jdbc.dcn;
 
 /**
- * Mocking class
+ * A mocking class used as a placeholder for the real Oracle JDBC drivers to prevent build errors.
  */
 public class DatabaseChangeEvent {
     // no data
