+++ src/oracleMock/oracle/jdbc/driver/OracleDriver.java
@@ -1,7 +1,7 @@
 package oracle.jdbc.driver;
 
 /**
- * Mocking class
+ * A mocking class used as a placeholder for the real Oracle JDBC drivers to prevent build errors.
  */
 public class OracleDriver {
     // no data
